added reset password for agents, rearranged files in the auth folder and added button to redirect after password is reset
