+++ src/App.js
@@ -1,205 +1,193 @@
-import { ThemeProvider, createTheme } from "@mui/material";
-import { useContext } from "react";
-import { CookiesProvider } from "react-cookie";
-import { Navigate, Route, Routes, useLocation, useNavigate } from "react-router-dom";
-import "./App.css";
-import ProtectedRoute from "./components/protected-route";
-import UserProtectedRoute from "./components/user-protected-route";
-import { AuthContext } from "./context/AuthContext";
-import { Agents } from "./pages/agent/Agents";
-import { AgentSignIn } from "./pages/auth/agent-sign-in";
-import { UserEmailConfirmation } from "./pages/auth/user-email-confirmation";
-import { AgentConfirmation } from "./pages/auth/agent-confirmation";
-import { UserResetPassword } from "./pages/auth/user-reset-password";
-import { UserResetPasswordConfirmation } from "./pages/auth/user-reset-password-confirmation";
-import { UserResetPasswordForm } from "./pages/auth/user-reset-password-form";
-import { UserSignIn } from "./pages/auth/user-sign-in";
-import { UserSignUp } from "./pages/auth/user-sign-up";
-import { UserSignUpConfirmation } from "./pages/auth/user-sign-up-confirmation";
-import { AgentDashboard } from "./pages/dashboard/AgentDashboard";
-import { UserTickets } from "./pages/dashboard/UserTickets";
-import { Departments } from "./pages/department/Departments";
-import { EmailBanlist } from "./pages/email/EmailBanlist";
-import { EmailDiagnostic } from "./pages/email/EmailDiagnostic";
-import { Emails } from "./pages/email/emails/Emails";
-import { EmailSettings } from "./pages/email/EmailSettings";
-import { EmailTemplates } from "./pages/email/templates/EmailTemplates";
-import { Forms } from "./pages/form/Forms";
-import { Groups } from "./pages/group/Groups";
-import { Landing } from "./pages/landing/landing";
-import { Profile } from "./pages/profile/AgentProfile";
-import { Roles } from "./pages/role/Roles";
-import { Settings } from "./pages/settings/Settings";
-import {
-	AgentMenu,
-	CompanyMenu,
-	KnowledgebaseMenu,
-	SystemMenu,
-	TaskMenu,
-	TicketMenu,
-	UserMenu,
-} from "./pages/settings/SettingsMenus";
-import { Tickets } from "./pages/ticket/Tickets";
-import { Topics } from "./pages/topic/Topics";
-import { Users } from "./pages/user/Users";
-import { Sidebar } from "./components/sidebar";
-import { drawerWidth } from "./components/sidebar";
-import axios from "axios";
-import { TicketView } from "./pages/ticket/TicketView";
+import { ThemeProvider, createTheme } from '@mui/material';
+import axios from 'axios';
+import { useContext } from 'react';
+import { CookiesProvider } from 'react-cookie';
+import { Navigate, Route, Routes, useNavigate } from 'react-router-dom';
+import './App.css';
+import ProtectedRoute from './components/protected-route';
+import { Sidebar } from './components/sidebar';
+import UserProtectedRoute from './components/user-protected-route';
+import { AuthContext } from './context/AuthContext';
+import { Agents } from './pages/agent/Agents';
+import { AgentConfirmation } from './pages/auth/agent/agent-confirmation';
+import { AgentSignIn } from './pages/auth/agent/agent-sign-in';
+import { AgentResetPassword } from './pages/auth/agent/agent-reset-password';
+import { AgentResetPasswordConfirmation } from './pages/auth/agent/agent-reset-password-confirmation';
+import { AgentResetPasswordForm } from './pages/auth/agent/agent-reset-password-form';
+import { Users } from './pages/user/Users';
+import { UserSignUpConfirmation } from './pages/auth/user/user-sign-up-confirmation';
+import { UserEmailConfirmation } from './pages/auth/user/user-email-confirmation';
+import { UserSignIn } from './pages/auth/user/user-sign-in';
+import { UserSignUp } from './pages/auth/user/user-sign-up';
+import { UserResetPassword } from './pages/auth/user/user-reset-password';
+import { UserResetPasswordConfirmation } from './pages/auth/user/user-reset-password-confirmation';
+import { UserResetPasswordForm } from './pages/auth/user/user-reset-password-form';
+import { AgentDashboard } from './pages/dashboard/AgentDashboard';
+import { UserTickets } from './pages/dashboard/UserTickets';
+import { Departments } from './pages/department/Departments';
+import { EmailBanlist } from './pages/email/EmailBanlist';
+import { EmailDiagnostic } from './pages/email/EmailDiagnostic';
+import { Emails } from './pages/email/emails/Emails';
+import { EmailSettings } from './pages/email/EmailSettings';
+import { EmailTemplates } from './pages/email/templates/EmailTemplates';
+import { Forms } from './pages/form/Forms';
+import { Groups } from './pages/group/Groups';
+import { Landing } from './pages/landing/landing';
+import { Profile } from './pages/profile/AgentProfile';
+import { Roles } from './pages/role/Roles';
+import { Settings } from './pages/settings/Settings';
+import { AgentMenu, CompanyMenu, KnowledgebaseMenu, SystemMenu, TaskMenu, TicketMenu, UserMenu } from './pages/settings/SettingsMenus';
+import { Tickets } from './pages/ticket/Tickets';
+import { TicketView } from './pages/ticket/TicketView';
+import { Topics } from './pages/topic/Topics';
 
 const theme = createTheme({
 	palette: {
 		primary: {
-			main: "#22874E",
+			main: '#22874E',
 		},
 	},
 	typography: {
-		fontFamily: [
-			"Mont",
-			"Roboto",
-			'"Helvetica Neue"',
-			"Arial",
-			"sans-serif",
-		].join(","),
+		fontFamily: ['Mont', 'Roboto', '"Helvetica Neue"', 'Arial', 'sans-serif'].join(','),
 		h1: {
-			fontSize: "2rem",
+			fontSize: '2rem',
 			lineHeight: 1.1875,
 			fontWeight: 600,
-			letterSpacing: "-0.03em",
+			letterSpacing: '-0.03em',
 		},
 		h2: {
-			fontSize: "1.5rem",
+			fontSize: '1.5rem',
 			fontWeight: 600,
-			letterSpacing: "-0.02em",
+			letterSpacing: '-0.02em',
 		},
 		h3: {
-			fontSize: "1.25rem",
+			fontSize: '1.25rem',
 			fontWeight: 500,
-			letterSpacing: "-0.02em",
+			letterSpacing: '-0.02em',
 		},
 		h4: {
-			fontSize: "1.125rem",
+			fontSize: '1.125rem',
 			fontWeight: 500,
-			letterSpacing: "-0.02em",
+			letterSpacing: '-0.02em',
 		},
 		h6: {
-			letterSpacing: "-0.02em",
+			letterSpacing: '-0.02em',
 		},
 		subtitle1: {
-			letterSpacing: "-0.05em",
+			letterSpacing: '-0.05em',
 		},
 		subtitle2: {
-			color: "#545555",
-			letterSpacing: "-0.05em",
+			color: '#545555',
+			letterSpacing: '-0.05em',
 		},
 		caption: {
-			letterSpacing: "-0.03em",
-			fontFamily: [
-				"Mont",
-				"Roboto",
-				'"Helvetica Neue"',
-				"Arial",
-				"sans-serif",
-			].join(","),
+			letterSpacing: '-0.03em',
+			fontFamily: ['Mont', 'Roboto', '"Helvetica Neue"', 'Arial', 'sans-serif'].join(','),
 		},
 		overline: {
 			fontWeight: 600,
-			lineHeight: "unset",
-			letterSpacing: "-0.03em",
+			lineHeight: 'unset',
+			letterSpacing: '-0.03em',
 		},
 	},
 });
 
 function App() {
 	const { agentAuthState, userAuthState, agentLogout, userLogout } = useContext(AuthContext);
-	const navigate = useNavigate()
+	const navigate = useNavigate();
 
-	axios.interceptors.response.use(response => {
-		return response;
-	}, error => {
-		if (error.response.status === 401) {
-			agentLogout()
-			userLogout()
-			navigate("/")
-			return error
+	axios.interceptors.response.use(
+		(response) => {
+			return response;
+		},
+		(error) => {
+			if (error.response.status === 401) {
+				agentLogout();
+				userLogout();
+				navigate('/');
+				return error;
+			}
+			throw error;
 		}
-		throw error;
-	});
+	);
 
 	return (
 		<ThemeProvider theme={theme}>
-			<CookiesProvider defaultSetOptions={{ path: "/" }}>
-				<div className="App">
+			<CookiesProvider defaultSetOptions={{ path: '/' }}>
+				<div className='App'>
 					<Routes>
-						<Route path="/" exact element={<Landing />} />
-						<Route
-							path="agent/login"
-							element={
-								agentAuthState.isAuth ? (
-									<Navigate to="/tickets" />
-								) : (
-									<AgentSignIn />
-								)
-							}
+						<Route 
+							path='/' 
+							exact element={<Landing />} 
+						/>
+						<Route 
+							path='agent/login' 
+							element={agentAuthState.isAuth ? <Navigate to='/tickets' /> : <AgentSignIn />} 
+						/>
+						<Route 
+							path='user/login' 
+							element={userAuthState.isAuth ? <Navigate to='/user/tickets' /> : <UserSignIn />} 
+						/>
+						<Route 
+							path='signup' 
+							element={<UserSignUp />} 
+						/>
+						<Route 
+							path='signup/confirmation/:user_id' 
+							element={<UserSignUpConfirmation />} 
+						/>
+						<Route 
+							path='/reset_password' 
+							element={<UserResetPassword />} 
+						/>
+						<Route 
+							path='/reset_password/confirmation/:user_id' 
+							element={<UserResetPasswordConfirmation />} 
+						/>
+						<Route 
+							path='/reset_password/:token' 
+							element={<UserResetPasswordForm />} 
 						/>
 						<Route
-							path="user/login"
-							element={
-								userAuthState.isAuth ? (
-									<Navigate to="/user/tickets" />
-								) : (
-									<UserSignIn />
-								)
-							}
-						/>
-						<Route path="signup" element={<UserSignUp />} />
-						<Route
-							path="signup/confirmation/:user_id"
-							element={<UserSignUpConfirmation />}
+							path="/agent/reset_password"
+							element={<AgentResetPassword />}
 						/>
 						<Route
-							path="/reset_password"
-							element={<UserResetPassword />}
+							path="/agent/reset_password/confirmation/:agent_id"
+							element={<AgentResetPasswordConfirmation />}
 						/>
 						<Route
-							path="/reset_password/confirmation/:user_id"
-							element={<UserResetPasswordConfirmation />}
+							path="/agent/reset_password/:token"
+							element={<AgentResetPasswordForm />}
+						/>
+						<Route 
+							path='confirm_email/:token' 
+							element={<UserEmailConfirmation />} 
+						/>
+						<Route 
+							path='confirm_agent_email/:token' 
+							element={<AgentConfirmation />} 
 						/>
 						<Route
-							path="/reset_password/:token"
-							element={<UserResetPasswordForm />}
-						/>
-						<Route
-							path="confirm_email/:token"
-							element={<UserEmailConfirmation />}
-						/>
-						<Route
-							path="confirm_agent_email/:token"
-							element={<AgentConfirmation />}
-						/>
-						<Route
-								path="user/tickets/:ticketId"
-								element={
-									<UserProtectedRoute>
-										<UserTickets />
-									</UserProtectedRoute>
-								}
-							/>
-						<Route
-							path="user/tickets"
+							path='user/tickets/:ticketId'
 							element={
 								<UserProtectedRoute>
 									<UserTickets />
 								</UserProtectedRoute>
 							}
 						/>
-						<Route element={<Sidebar/>}>
+						<Route
+							path='user/tickets'
+							element={
+								<UserProtectedRoute>
+									<UserTickets />
+								</UserProtectedRoute>
+							}
+						/>
+						<Route element={<Sidebar />}>
+							<Route path='dashboard' element={<AgentDashboard />} />
 							<Route
-								path="dashboard"
-								element={<AgentDashboard />}
-							/>
-							<Route
-								path="tickets"
+								path='tickets'
 								element={
 									<ProtectedRoute>
 										<Tickets />
@@ -207,7 +195,7 @@ function App() {
 								}
 							/>
 							<Route
-								path="tickets/id/:ticketId"
+								path='tickets/id/:ticketId'
 								element={
 									<ProtectedRoute>
 										<Tickets />
@@ -215,7 +203,7 @@ function App() {
 								}
 							/>
 							<Route
-								path="tickets/:number"
+								path='tickets/:number'
 								element={
 									<ProtectedRoute>
 										<TicketView />
@@ -223,7 +211,7 @@ function App() {
 								}
 							/>
 							<Route
-								path="settings/system"
+								path='settings/system'
 								element={
 									<ProtectedRoute requireAdmin>
 										<Settings Menu={SystemMenu} />
@@ -231,7 +219,7 @@ function App() {
 								}
 							/>
 							<Route
-								path="settings/company"
+								path='settings/company'
 								element={
 									<ProtectedRoute requireAdmin>
 										<Settings Menu={CompanyMenu} />
@@ -239,7 +227,7 @@ function App() {
 								}
 							/>
 							<Route
-								path="settings/tickets"
+								path='settings/tickets'
 								element={
 									<ProtectedRoute requireAdmin>
 										<Settings Menu={TicketMenu} />
@@ -247,7 +235,7 @@ function App() {
 								}
 							/>
 							<Route
-								path="settings/tasks"
+								path='settings/tasks'
 								element={
 									<ProtectedRoute requireAdmin>
 										<Settings Menu={TaskMenu} />
@@ -255,7 +243,7 @@ function App() {
 								}
 							/>
 							<Route
-								path="settings/agents"
+								path='settings/agents'
 								element={
 									<ProtectedRoute requireAdmin>
 										<Settings Menu={AgentMenu} />
@@ -263,7 +251,7 @@ function App() {
 								}
 							/>
 							<Route
-								path="settings/users"
+								path='settings/users'
 								element={
 									<ProtectedRoute requireAdmin>
 										<Settings Menu={UserMenu} />
@@ -271,7 +259,7 @@ function App() {
 								}
 							/>
 							<Route
-								path="settings/knowledgebase"
+								path='settings/knowledgebase'
 								element={
 									<ProtectedRoute requireAdmin>
 										<Settings Menu={KnowledgebaseMenu} />
@@ -280,7 +268,7 @@ function App() {
 							/>
 
 							<Route
-								path="email/emails"
+								path='email/emails'
 								element={
 									<ProtectedRoute requireAdmin>
 										<Emails />
@@ -289,7 +277,7 @@ function App() {
 							/>
 
 							<Route
-								path="email/settings"
+								path='email/settings'
 								element={
 									<ProtectedRoute requireAdmin>
 										<EmailSettings />
@@ -298,7 +286,7 @@ function App() {
 							/>
 
 							<Route
-								path="email/banlist"
+								path='email/banlist'
 								element={
 									<ProtectedRoute requireAdmin>
 										<EmailBanlist />
@@ -307,7 +295,7 @@ function App() {
 							/>
 
 							<Route
-								path="email/templates"
+								path='email/templates'
 								element={
 									<ProtectedRoute requireAdmin>
 										<EmailTemplates />
@@ -316,7 +304,7 @@ function App() {
 							/>
 
 							<Route
-								path="email/templates/:templateId"
+								path='email/templates/:templateId'
 								element={
 									<ProtectedRoute requireAdmin>
 										<EmailTemplates />
@@ -325,7 +313,7 @@ function App() {
 							/>
 
 							<Route
-								path="email/diagnostic"
+								path='email/diagnostic'
 								element={
 									<ProtectedRoute requireAdmin>
 										<EmailDiagnostic />
@@ -334,7 +322,7 @@ function App() {
 							/>
 
 							<Route
-								path="profile"
+								path='profile'
 								element={
 									<ProtectedRoute>
 										<Profile />
@@ -343,29 +331,25 @@ function App() {
 							/>
 
 							<Route
-								path="manage/agents"
+								path='manage/agents'
 								element={
-									<ProtectedRoute
-										requirePermission={"agent.view"}
-									>
+									<ProtectedRoute requirePermission={'agent.view'}>
 										<Agents />
 									</ProtectedRoute>
 								}
 							/>
 
 							<Route
-								path="manage/groups"
+								path='manage/groups'
 								element={
-									<ProtectedRoute
-										requirePermission={"agent.view"}
-									>
+									<ProtectedRoute requirePermission={'agent.view'}>
 										<Groups />
 									</ProtectedRoute>
 								}
 							/>
 
 							<Route
-								path="manage/users"
+								path='manage/users'
 								element={
 									<ProtectedRoute>
 										<Users />
@@ -374,7 +358,7 @@ function App() {
 							/>
 
 							<Route
-								path="manage/queues"
+								path='manage/queues'
 								element={
 									<ProtectedRoute>
 										<Users />
@@ -383,7 +367,7 @@ function App() {
 							/>
 
 							<Route
-								path="manage/slas"
+								path='manage/slas'
 								element={
 									<ProtectedRoute>
 										<Users />
@@ -392,7 +376,7 @@ function App() {
 							/>
 
 							<Route
-								path="manage/schedules"
+								path='manage/schedules'
 								element={
 									<ProtectedRoute>
 										<Users />
@@ -401,7 +385,7 @@ function App() {
 							/>
 
 							<Route
-								path="manage/departments"
+								path='manage/departments'
 								element={
 									<ProtectedRoute>
 										<Departments />
@@ -410,7 +394,7 @@ function App() {
 							/>
 
 							<Route
-								path="manage/groups"
+								path='manage/groups'
 								element={
 									<ProtectedRoute>
 										<Groups />
@@ -418,7 +402,7 @@ function App() {
 								}
 							/>
 							<Route
-								path="manage/topics"
+								path='manage/topics'
 								element={
 									<ProtectedRoute>
 										<Topics />
@@ -426,7 +410,7 @@ function App() {
 								}
 							/>
 							<Route
-								path="manage/forms"
+								path='manage/forms'
 								element={
 									<ProtectedRoute>
 										<Forms />
@@ -434,7 +418,7 @@ function App() {
 								}
 							/>
 							<Route
-								path="manage/roles"
+								path='manage/roles'
 								element={
 									<ProtectedRoute>
 										<Roles />
@@ -442,7 +426,6 @@ function App() {
 								}
 							/>
 						</Route>
-
 					</Routes>
 				</div>
 			</CookiesProvider>
