+++ src/hooks/useAgentBackend.js
@@ -5,6 +5,18 @@ import { AuthContext } from '../context/AuthContext';
 export const useAgentBackend = () => {
 	const { agentAuthState } = useContext(AuthContext);
 
+	const sendResetPasswordEmail = async (email) => {
+		return await axios.post(process.env.REACT_APP_BACKEND_URL + `agent/reset`, {'email': email});
+	}
+
+	const resendResetPasswordEmail = async (agent_id) => {
+		return await axios.post(process.env.REACT_APP_BACKEND_URL + `agent/reset/resend/${agent_id}`);
+	}
+
+	const confirmResetPasswordToken = async (token, password) => {
+		return await axios.post(process.env.REACT_APP_BACKEND_URL + `agent/reset/confirm/${token}`, {'password': password});
+	}
+	
 	const getAllAgents = async () => {
 		const config = {
 			headers: { Authorization: `Bearer ${agentAuthState.token}` },
@@ -102,5 +114,5 @@ export const useAgentBackend = () => {
 		);
 	};
 
-	return { getAllAgents, getAllAgentsByDeptAndGroup, getPermissions, getAgentBySearch, createAgent, updateAgent, removeAgent, getAgentById, confirmToken, registerAgent, resendConfirmationEmail };
+	return { getAllAgents, getAllAgentsByDeptAndGroup, getPermissions, getAgentBySearch, createAgent, updateAgent, removeAgent, getAgentById, confirmToken, registerAgent, resendConfirmationEmail, sendResetPasswordEmail, resendResetPasswordEmail, confirmResetPasswordToken };
 };
