+++ src/pages/auth/agent/agent-reset-password-confirmation.jsx
@@ -1,24 +1,17 @@
-import React, { useContext, useState } from 'react';
-import '../../App.css';
-import AppIcon from '../../assets/app-icon-black.png';
-import logoBlack from '../../assets/logo-black.svg';
-import logo from '../../assets/logo-white.svg';
-
 import {
 	Box,
-	Button,
-	CircularProgress,
-	InputAdornment,
-	Link,
-	TextField,
+	Button, TextField,
 	Typography,
-	styled,
+	styled
 } from '@mui/material';
 import Grid from '@mui/material/Grid2';
-import { Activity, Lock, Mail, Split, Tag } from 'lucide-react';
-import { useNavigate } from 'react-router-dom';
-import { AuthContext } from '../../context/AuthContext';
-import { useSetAuthCookie } from '../../hooks/useSetAuthCookie';
+import { Activity, CheckCircle, Split, Tag } from 'lucide-react';
+import React from 'react';
+import { useParams } from 'react-router-dom';
+import '../../../App.css';
+import logoBlack from '../../../assets/logo-black.svg';
+import logo from '../../../assets/logo-white.svg';
+import { useAgentBackend } from '../../../hooks/useAgentBackend';
 
 const ProviderButton = styled(Box)({
 	border: '2px solid #EFEFEF',
@@ -73,61 +66,18 @@ const RedirectButton = styled('a')({
 	},
 });
 
-export const UserSignIn = () => {
-	const { setUserData } = useContext(AuthContext);
-	const [loading, setLoading] = useState(false);
+export const AgentResetPasswordConfirmation = () => {
 
-	const [email, setEmail] = useState('');
-	const [error, setError] = useState(false);
+	const { agent_id } = useParams()
+	const { resendResetPasswordEmail } = useAgentBackend()
 
-	const [password, setPassword] = useState('');
-	const [passwordError, setPasswordError] = useState(false);
-
-	const navigate = useNavigate();
-	const { userSignInEmailAndPassword } = useSetAuthCookie();
-
-	const signIn = async e => {
-		e.preventDefault();
-		setLoading(true);
-
-		if (validateEmail(email) && password !== '') {
-			userSignInEmailAndPassword(email, password)
-				// signInWithEmailAndPassword(auth, email, password)
-				.then(userCredential => {
-					// getApiToken(userCredential);
-					const userData = userCredential.data;
-
-					const authInfo = {
-						isAuth: true,
-						user_id: userData.user_id,
-						token: userData.token,
-					};
-					setUserData(authInfo);
-					setLoading(false);
-					navigate('/user/tickets');
-				})
-				.catch(error => {
-					const errorCode = error.code;
-					const errorMessage = error.message;
-					console.error(errorCode, errorMessage);
-					setLoading(false);
-				});
-		} else if (!validateEmail(email)) {
-			setError(true);
-			setLoading(false);
-		} else if (password === '') {
-			setPasswordError(true);
-			setLoading(false);
-		}
-	};
-
-	const validateEmail = email => {
-		return String(email)
-			.toLowerCase()
-			.match(
-				/^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|.(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/
-			);
-	};
+	const resendEmail = () => {
+		resendResetPasswordEmail(agent_id)
+		.catch(error => {
+			console.error(error);
+		});
+		window.location.reload();
+	}
 
 	return (
 		<Box
@@ -192,6 +142,57 @@ export const UserSignIn = () => {
 							>
 								Experience the future of customer support with Triage.ai
 							</Typography>
+
+							{/* <Box
+								sx={{
+									display: 'flex',
+									flexDirection: 'column',
+									gap: '22px',
+									// textAlign: 'left',
+									fontSize: '0.875rem',
+									color: '#7A8087',
+								}}
+							>
+								<div>
+									<span style={{ display: 'inline-block', fontWeight: '600' }}>
+										Build, Fine-Tune, Test, and Deploy your own ticket classification system in a few
+										clicks!
+									</span>
+								</div>
+
+								<Box sx={{ display: 'flex', alignItems: 'flex-start', textAlign: 'left' }}>
+									<CheckCircle
+										color="#8CC279"
+										size={22}
+										style={{ flexShrink: 0 }}
+									/>
+									<span style={{ fontWeight: '500', marginLeft: '12px', marginTop: '2px' }}>
+										Auto-labels tickets
+									</span>
+								</Box>
+
+								<Box sx={{ display: 'flex', alignItems: 'flex-start', textAlign: 'left' }}>
+									<CheckCircle
+										color="#8CC279"
+										size={22}
+										style={{ flexShrink: 0 }}
+									/>
+									<span style={{ fontWeight: '500', marginLeft: '12px', marginTop: '2px' }}>
+										Ensures accurate ticket assignment
+									</span>
+								</Box>
+
+								<Box sx={{ display: 'flex', alignItems: 'flex-start', textAlign: 'left' }}>
+									<CheckCircle
+										color="#8CC279"
+										size={22}
+										style={{ flexShrink: 0 }}
+									/>
+									<span style={{ fontWeight: '500', marginLeft: '12px', marginTop: '2px' }}>
+										Pinpoints areas experiencing a surge in ticket volume
+									</span>
+								</Box>
+							</Box> */}
 
 							<Grid
 								container
@@ -293,7 +294,7 @@ export const UserSignIn = () => {
 							// backgroundColor: '#FCFCFC',
 						}}
 					>
-						<header className="App-header">
+						<header className="App-header-wide">
 							<Box
 								sx={{
 									width: '100%',
@@ -331,169 +332,49 @@ export const UserSignIn = () => {
 								</Typography> */}
 							</Box>
 
-							<img
+							{/* <img
 								src={AppIcon}
 								className="App-logo"
 								// style={{ width: '0px' }}
 								alt="logo"
-							/>
+							/> */}
 
-							<h1
+							<CheckCircle size={60} color='#34b233' />
+							<h5
 								style={{
-									fontSize: '3rem',
-									fontWeight: 600,
-									color: '#1B1D1F',
-									letterSpacing: '-0.03em',
-									marginTop: '30px',
-									marginBottom: '30px',
-								}}
-							>
-								User Sign in
-							</h1>
-
-							{/* <p
-								style={{
-									fontSize: '0.875rem',
+									// fontSize: '0.95rem',
 									fontWeight: 600,
 									color: '#1B1D1F',
 									letterSpacing: '-0.01em',
 									lineHeight: 1.2,
-									marginTop: 0,
-									marginBottom: '20px',
+									marginTop: '20px',
+									marginBottom: 0,
 									textAlign: 'center',
 								}}
 							>
-								Sign in with a provider
-							</p>
-
-							<Box sx={{ display: 'flex', width: '100%', gap: '10px', mb: '35px' }}>
-								<ProviderButton onClick={loginWithSAML}>
-									<img
-										src={microsoftIcon}
-										alt="Microsoft Icon"
-									/>
-									<span
-										style={{
-											fontSize: '0.9375rem',
-											fontWeight: 700,
-											color: '#1B1D1F',
-											marginLeft: '8px',
-										}}
-									>
-										Microsoft
-									</span>
-								</ProviderButton>
-
-								<ProviderButton>
-									<img
-										src={googleIcon}
-										alt="Google Icon"
-									/>
-									<span
-										style={{
-											fontSize: '0.9375rem',
-											fontWeight: 700,
-											color: '#1B1D1F',
-											marginLeft: '8px',
-										}}
-									>
-										Google
-									</span>
-								</ProviderButton>
-							</Box>
-
-							<hr style={{ width: '100%', border: '1px solid #EFEFEF', margin: 0 }} /> */}
-
-							{/* <span
-								style={{
-									fontSize: '0.875rem',
-									fontWeight: 600,
-									color: '#1B1D1F',
-									letterSpacing: '-0.01em',
-									lineHeight: 1.2,
-									marginTop: '32px',
-									marginBottom: '25px',
-								}}
-							>
-								Or continue with email and password
-							</span> */}
+								Check Your Email
+							</h5>
 
 							<p
 								style={{
-									fontSize: '0.875rem',
+									fontSize: '0.8rem',
 									fontWeight: 600,
 									color: '#1B1D1F',
 									letterSpacing: '-0.01em',
 									lineHeight: 1.2,
-									marginTop: 0,
+									marginTop: '10px',
 									marginBottom: '20px',
 									textAlign: 'center',
 								}}
 							>
-								Sign in with email and password
+								Follow the steps in the confirmation sent to your email
 							</p>
 
-							<form onSubmit={e => signIn(e)}>
-								<CustomTextField
-									label=""
-									id="email"
-									autoComplete="username"
-									sx={{
-										mb: 1,
-										'& .MuiInputBase-root': {
-											border: error ? '2px solid #ff7474' : '2px solid transparent',
-										},
-									}}
-									placeholder="Your email"
-									InputProps={{
-										startAdornment: (
-											<InputAdornment position="start">
-												<Mail color="#575757" />
-											</InputAdornment>
-										),
-									}}
-									value={email}
-									onChange={event => {
-										if (validateEmail(email)) {
-											setError(false);
-										}
-										setEmail(event.target.value);
-									}}
-								/>
-
-								<CustomTextField
-									label=""
-									id="password"
-									type="password"
-									autoComplete="current-password"
-									sx={{
-										'& .MuiInputBase-root': {
-											border: passwordError ? '2px solid #ff7474' : '2px solid transparent',
-										},
-									}}
-									placeholder="Your password"
-									InputProps={{
-										startAdornment: (
-											<InputAdornment position="start">
-												<Lock color="#575757" />
-											</InputAdornment>
-										),
-									}}
-									value={password}
-									onChange={event => {
-										// if (validatePassword(password)) {
-											// 	setError(false);
-											// }
-											setPassword(event.target.value);
-										}}
-								/>
-
-								<Button
+							<Button
 									sx={{
 										backgroundColor: '#22874E',
 										color: '#FFF',
 										borderRadius: '12px',
-										width: '100%',
 										fontSize: '0.9375rem',
 										fontWeight: 600,
 										lineHeight: 1,
@@ -510,62 +391,12 @@ export const UserSignIn = () => {
 											opacity: 0.4,
 										},
 									}}
-									type="submit"
-									disabled={loading || !validateEmail(email) || password === ''}
+									onClick={() => resendEmail()}
 								>
-									{loading ? (
-										<CircularProgress
-											size={22}
-											thickness={5}
-											sx={{ color: '#FFF' }}
-										/>
-									) : (
-										'Sign in'
-									)}
+									Resend Email
 								</Button>
-							</form>
-							<p
-								style={{
-									fontSize: '0.875rem',
-									fontWeight: 600,
-									color: '#1B1D1F',
-									letterSpacing: '-0.01em',
-									lineHeight: 1.2,
-									marginTop: 5,
-									marginBottom: '20px',
-									textAlign: 'center',
-								}}
-							>
-										<Link underline='none' component='button' onClick={() => navigate('/reset_password')}>Forgot password?</Link>
-							</p>
-						<p
-							style={{
-								fontSize: '0.875rem',
-								fontWeight: 600,
-								color: '#1B1D1F',
-								letterSpacing: '-0.01em',
-								lineHeight: 1.2,
-								marginTop: 0,
-								marginBottom: '10px',
-								textAlign: 'center',
-							}}
-						>
-							Don't have an account? Sign up <Link underline='none' component='button' onClick={() => navigate('/signup')}>here</Link>
-						</p>
-						<p
-							style={{
-								fontSize: '0.875rem',
-								fontWeight: 600,
-								color: '#1B1D1F',
-								letterSpacing: '-0.01em',
-								lineHeight: 1.2,
-								marginTop: 0,
-								marginBottom: '20px',
-								textAlign: 'center',
-							}}
-						>
-							Looking for agent sign in? <Link underline='none' component='button' onClick={() => navigate('/agent/login')}>Click here</Link>
-						</p>
+
+						
 						</header>
 					</div>
 				</Grid>
