+++ src/pages/auth/user/user-reset-password.jsx
@@ -1,17 +1,12 @@
-import {
-	Box,
-	Button, TextField,
-	Typography,
-	styled
-} from '@mui/material';
+import { Box, Button, InputAdornment, TextField, Typography, styled } from '@mui/material';
 import Grid from '@mui/material/Grid2';
-import { Activity, CheckCircle, Split, Tag } from 'lucide-react';
-import React from 'react';
-import { useParams } from 'react-router-dom';
-import '../../App.css';
-import logoBlack from '../../assets/logo-black.svg';
-import logo from '../../assets/logo-white.svg';
-import { useUserBackend } from '../../hooks/useUserBackend';
+import { Activity, Mail, Split, Tag } from 'lucide-react';
+import React, { useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { useUserBackend } from '../../../hooks/useUserBackend';
+import '../../../App.css';
+import logoBlack from '../../../assets/logo-black.svg';
+import logo from '../../../assets/logo-white.svg';
 
 const ProviderButton = styled(Box)({
 	border: '2px solid #EFEFEF',
@@ -66,18 +61,29 @@ const RedirectButton = styled('a')({
 	},
 });
 
-export const UserSignUpConfirmation = () => {
+export const UserResetPassword = () => {
+	const { sendResetPasswordEmail } = useUserBackend();
 
-	const { user_id } = useParams()
-	const { resendConfirmationEmail } = useUserBackend()
+	const [email, setEmail] = useState('');
+	const [error, setError] = useState(false);
 
-	const resendEmail = () => {
-		resendConfirmationEmail(user_id)
-		.catch(error => {
-			console.error(error);
-		});
-		window.location.reload();
-	}
+	const navigate = useNavigate();
+
+	const validateEmail = (email) => {
+		return String(email)
+			.toLowerCase()
+			.match(
+				/^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|.(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/
+			);
+	};
+
+	const sendResetPassword = () => {
+		sendResetPasswordEmail(email)
+			.then((res) => {
+				navigate('/reset_password/confirmation/' + res.data.user_id);
+			})
+			.catch((err) => console.error(err));
+	};
 
 	return (
 		<Box
@@ -89,10 +95,7 @@ export const UserSignUpConfirmation = () => {
 				backgroundColor: '#FCFCFC',
 			}}
 		>
-			<Grid
-				container
-				spacing={{ xs: 6, md: 8, lg: 2 }}
-			>
+			<Grid container spacing={{ xs: 6, md: 8, lg: 2 }}>
 				<Grid
 					size={{ xs: 0, md: 6 }}
 					sx={{
@@ -111,8 +114,7 @@ export const UserSignUpConfirmation = () => {
 							sx={{
 								width: '100%',
 								height: '100%',
-								background:
-									'radial-gradient(130% 135% at 30% 10%, #0000 40%, #0da54d, #D0FFD6), #010312',
+								background: 'radial-gradient(130% 135% at 30% 10%, #0000 40%, #0da54d, #D0FFD6), #010312',
 								display: { xs: 'none', md: 'flex' },
 								flexDirection: 'column',
 								alignItems: 'flex-start',
@@ -125,7 +127,7 @@ export const UserSignUpConfirmation = () => {
 							}}
 						>
 							<Typography
-								variant="h1"
+								variant='h1'
 								sx={{
 									fontSize: '3.75rem',
 									background: 'radial-gradient(45% 100% at 50% 50%, #fff 50%, #ffffff80)',
@@ -194,88 +196,42 @@ export const UserSignUpConfirmation = () => {
 								</Box>
 							</Box> */}
 
-							<Grid
-								container
-								spacing={2}
-								sx={{ marginTop: '1.5rem', display: { xs: 'none', lg: 'flex' } }}
-							>
-								<Grid
-									size={{ xs: 4 }}
-									sx={{ textAlign: 'left' }}
-								>
+							<Grid container spacing={2} sx={{ marginTop: '1.5rem', display: { xs: 'none', lg: 'flex' } }}>
+								<Grid size={{ xs: 4 }} sx={{ textAlign: 'left' }}>
 									<Box sx={{ display: 'flex', alignItems: 'center', marginBottom: '0.2rem' }}>
-										<Tag
-											color="#fff"
-											size={16}
-											strokeWidth={2.2}
-											style={{ opacity: 0.6, marginRight: '0.4rem' }}
-										/>
-										<Typography
-											variant="subtitle1"
-											sx={{ color: '#FFF', lineHeight: 1.25, fontWeight: 500 }}
-										>
+										<Tag color='#fff' size={16} strokeWidth={2.2} style={{ opacity: 0.6, marginRight: '0.4rem' }} />
+										<Typography variant='subtitle1' sx={{ color: '#FFF', lineHeight: 1.25, fontWeight: 500 }}>
 											Labels tickets
 										</Typography>
 									</Box>
 
-									<Typography
-										variant="body2"
-										sx={{ color: '#FFF', opacity: 0.6 }}
-									>
+									<Typography variant='body2' sx={{ color: '#FFF', opacity: 0.6 }}>
 										Triage AI automatically labels your tickets to streamline your support process
 									</Typography>
 								</Grid>
 
-								<Grid
-									size={{ xs: 4 }}
-									sx={{ textAlign: 'left' }}
-								>
+								<Grid size={{ xs: 4 }} sx={{ textAlign: 'left' }}>
 									<Box sx={{ display: 'flex', alignItems: 'center', marginBottom: '0.2rem' }}>
-										<Split
-											color="#fff"
-											size={16}
-											strokeWidth={2.2}
-											style={{ opacity: 0.6, marginRight: '0.4rem' }}
-										/>
-										<Typography
-											variant="subtitle1"
-											sx={{ color: '#FFF', lineHeight: 1.25, fontWeight: 500 }}
-										>
+										<Split color='#fff' size={16} strokeWidth={2.2} style={{ opacity: 0.6, marginRight: '0.4rem' }} />
+										<Typography variant='subtitle1' sx={{ color: '#FFF', lineHeight: 1.25, fontWeight: 500 }}>
 											Assigns tickets
 										</Typography>
 									</Box>
 
-									<Typography
-										variant="body2"
-										sx={{ color: '#FFF', opacity: 0.6 }}
-									>
+									<Typography variant='body2' sx={{ color: '#FFF', opacity: 0.6 }}>
 										Ensures that tickets are accurately assigned to the appropriate members
 									</Typography>
 								</Grid>
 
-								<Grid
-									size={{ xs: 4 }}
-									sx={{ textAlign: 'left' }}
-								>
+								<Grid size={{ xs: 4 }} sx={{ textAlign: 'left' }}>
 									<Box sx={{ display: 'flex', alignItems: 'center', marginBottom: '0.2rem' }}>
-										<Activity
-											color="#fff"
-											size={16}
-											strokeWidth={2.2}
-											style={{ opacity: 0.6, marginRight: '0.4rem' }}
-										/>
-										<Typography
-											variant="subtitle1"
-											sx={{ color: '#FFF', lineHeight: 1.25, fontWeight: 500 }}
-										>
+										<Activity color='#fff' size={16} strokeWidth={2.2} style={{ opacity: 0.6, marginRight: '0.4rem' }} />
+										<Typography variant='subtitle1' sx={{ color: '#FFF', lineHeight: 1.25, fontWeight: 500 }}>
 											Identifies surges
 										</Typography>
 									</Box>
 
-									<Typography
-										variant="body2"
-										sx={{ color: '#FFF', opacity: 0.6 }}
-									>
+									<Typography variant='body2' sx={{ color: '#FFF', opacity: 0.6 }}>
 										Pinpoints areas with an increased ticket activity for proactive management
 									</Typography>
 								</Grid>
@@ -294,7 +250,7 @@ export const UserSignUpConfirmation = () => {
 							// backgroundColor: '#FCFCFC',
 						}}
 					>
-						<header className="App-header-wide">
+						<header className='App-header'>
 							<Box
 								sx={{
 									width: '100%',
@@ -309,19 +265,11 @@ export const UserSignUpConfirmation = () => {
 								}}
 							>
 								<Box sx={{ display: { xs: 'none', md: 'block' } }}>
-									<img
-										src={logo}
-										className="App-logo"
-										alt="logo"
-									/>
+									<img src={logo} className='App-logo' alt='logo' />
 								</Box>
 
 								<Box sx={{ display: { xs: 'block', md: 'none' } }}>
-									<img
-										src={logoBlack}
-										className="App-logo"
-										alt="logo"
-									/>
+									<img src={logoBlack} className='App-logo' alt='logo' />
 								</Box>
 
 								{/* <Typography
@@ -339,64 +287,87 @@ export const UserSignUpConfirmation = () => {
 								alt="logo"
 							/> */}
 
-							<CheckCircle size={60} color='#34b233' />
-							<h5
+							{/* <CheckCircle size={60} color='#34b233' /> */}
+
+							<h3
 								style={{
-									// fontSize: '0.95rem',
 									fontWeight: 600,
 									color: '#1B1D1F',
-									letterSpacing: '-0.01em',
-									lineHeight: 1.2,
-									marginTop: '20px',
-									marginBottom: 0,
-									textAlign: 'center',
+									letterSpacing: '-0.03em',
+									marginTop: '30px',
 								}}
 							>
-								Check Your Email
-							</h5>
-
+								Reset password
+							</h3>
 							<p
 								style={{
-									fontSize: '0.8rem',
+									fontSize: '0.875rem',
 									fontWeight: 600,
 									color: '#1B1D1F',
 									letterSpacing: '-0.01em',
 									lineHeight: 1.2,
-									marginTop: '10px',
+									marginTop: 0,
 									marginBottom: '20px',
 									textAlign: 'center',
 								}}
 							>
-								Follow the steps in the confirmation sent to your email
+								Enter your email below to reset your password
 							</p>
 
-							<Button
-									sx={{
-										backgroundColor: '#22874E',
-										color: '#FFF',
-										borderRadius: '12px',
-										fontSize: '0.9375rem',
-										fontWeight: 600,
-										lineHeight: 1,
-										textTransform: 'unset',
-										padding: '16.5px 10px',
-										marginTop: '12px',
-										marginBottom: '28px',
-										transition: 'all 0.3s',
-										'&:hover': {
-											backgroundColor: '#29b866',
-										},
-										'&:disabled': {
-											color: 'unset',
-											opacity: 0.4,
-										},
-									}}
-									onClick={() => resendEmail()}
-								>
-									Resend Email
-								</Button>
+							<CustomTextField
+								label=''
+								id='email'
+								autoComplete='username'
+								sx={{
+									mb: 1,
+									'& .MuiInputBase-root': {
+										border: error ? '2px solid #ff7474' : '2px solid transparent',
+									},
+								}}
+								placeholder='Your email'
+								InputProps={{
+									startAdornment: (
+										<InputAdornment position='start'>
+											<Mail color='#575757' />
+										</InputAdornment>
+									),
+								}}
+								value={email}
+								onChange={(event) => {
+									if (validateEmail(email)) {
+										setError(false);
+									}
+									setEmail(event.target.value);
+								}}
+							/>
 
-						
+							<Button
+								sx={{
+									backgroundColor: '#22874E',
+									color: '#FFF',
+									borderRadius: '12px',
+									width: '100%',
+									fontSize: '0.9375rem',
+									fontWeight: 600,
+									lineHeight: 1,
+									textTransform: 'unset',
+									padding: '16.5px 10px',
+									marginTop: '12px',
+									marginBottom: '28px',
+									transition: 'all 0.3s',
+									'&:hover': {
+										backgroundColor: '#29b866',
+									},
+									'&:disabled': {
+										color: 'unset',
+										opacity: 0.4,
+									},
+								}}
+								onClick={() => sendResetPassword()}
+								disabled={!validateEmail(email)}
+							>
+								Reset password
+							</Button>
 						</header>
 					</div>
 				</Grid>
