+++ src/pages/ticket/TicketThread.jsx
@@ -16,22 +16,20 @@ import {
 	Typography,
 } from '@mui/material';
 import { useEditor } from '@tiptap/react';
-import StarterKit from '@tiptap/starter-kit';
 import axios from 'axios';
 import dayjs from 'dayjs';
 import utc from 'dayjs/plugin/utc';
 import { CloudUploadIcon, File, Paperclip, Send, X } from 'lucide-react';
-import { RichTextReadOnly, TableBubbleMenu, TableImproved } from 'mui-tiptap';
+import { RichTextReadOnly } from 'mui-tiptap';
 import { useContext, useEffect, useState } from 'react';
 import { extensions, RichTextEditorBox } from '../../components/rich-text-editor';
 import { AuthContext } from '../../context/AuthContext';
 import formatDate from '../../functions/date-formatter';
-import  humanFileSize from '../../functions/file-size-formatter';
+import humanFileSize from '../../functions/file-size-formatter';
 import { useAttachmentBackend } from '../../hooks/useAttachmentsBackend';
-import { useThreadsBackend } from '../../hooks/useThreadBackend';
 import { useSettingsBackend } from '../../hooks/useSettingsBackend';
+import { useThreadsBackend } from '../../hooks/useThreadBackend';
 import { FileCard } from './FileCard';
-
 
 var localizedFormat = require('dayjs/plugin/localizedFormat');
 dayjs.extend(localizedFormat);
@@ -52,8 +50,8 @@ export const TicketThread = ({ ticket, closeDrawer, updateCurrentTicket, type })
 	const [formData, setFormData] = useState({ subject: null, body: '', type: 'A', editor: '', recipients: '' });
 	const [postDisabled, setPostDisabled] = useState(true);
 	const [files, setFiles] = useState([]);
-	const [maxUploadSize, setMaxUploadSize] = useState('')
-	const { createThreadEntry, createThreadEntryForUser } = useThreadsBackend();
+	const [maxUploadSize, setMaxUploadSize] = useState('');
+	const { createThreadEntry, createThreadEntryForUser, threadEntryAgentEmailReply, send_attachment_email } = useThreadsBackend();
 	const { getPresignedURL, createAttachment } = useAttachmentBackend();
 	const { getSettingsByKey } = useSettingsBackend();
 	const { permissions } = useContext(AuthContext);
@@ -69,9 +67,11 @@ export const TicketThread = ({ ticket, closeDrawer, updateCurrentTicket, type })
 		},
 	});
 
-	getSettingsByKey('agent_max_file_size').then(res => {
-		setMaxUploadSize(res.data.value)
-	})
+	useEffect(() => {
+		getSettingsByKey('agent_max_file_size').then((res) => {
+			setMaxUploadSize(res.data.value);
+		});
+	}, []);
 
 	const getDirection = (agent_id, option1, option2) => {
 		if ((agent_id && type === 'agent') || (!agent_id && type !== 'agent')) {
@@ -80,6 +80,7 @@ export const TicketThread = ({ ticket, closeDrawer, updateCurrentTicket, type })
 	};
 
 	const handleSubmit = async () => {
+		// ABSTRACT SO THAT I CAN USE IT FOR EITHER TYPE OF THREAD ENTRY
 		const threadEntryCreate = type === 'agent' ? createThreadEntry : createThreadEntryForUser;
 		var newThreadEntry = { ...formData, thread_id: ticket.thread.thread_id };
 		if (type === 'agent') {
@@ -89,69 +90,66 @@ export const TicketThread = ({ ticket, closeDrawer, updateCurrentTicket, type })
 		}
 
 		var updatedTicket = { ...ticket };
-		threadEntryCreate(newThreadEntry)
+		if (files.length !== 0) {
+			const file_names = files.map((item) => item.name);
+			
+			getPresignedURL({ attachment_names: file_names })
+			.then((res) => {
+				if (res.status !== 200) {
+					alert('Error with S3 configuration. Attachment will not be added to thread');
+					return {};
+				} else {
+					var presigned_urls = { ...res.data.url_dict };
+					return presigned_urls;
+				}
+			})
+			.then(async (presigned_urls) => {
+				if (Object.keys(presigned_urls).length === 0) {
+					return [];
+				} else {
+					var attachments = await awsFileUpload(presigned_urls, files);
+					return attachments;
+				}
+			})
+			.then(async (attachments) => {
+				newThreadEntry = attachments.length !== 0 ? { ...newThreadEntry, attachments: attachments } : newThreadEntry;
+				await threadEntryCreate(newThreadEntry)
+				.then((response) => {
+					updatedTicket.thread.entries.push(response.data);
+					editor.commands.setContent('');
+					setFormData({ subject: null, body: '', type: 'A', editor: '', recipients: '' });
+					setFiles([]);
+				})
+			})
+			.then(() => {
+				updateCurrentTicket(updatedTicket);
+			});
+		} else {
+			threadEntryCreate(newThreadEntry)
 			.then((response) => {
 				updatedTicket.thread.entries.push(response.data);
 				editor.commands.setContent('');
 				setFormData({ subject: null, body: '', type: 'A', editor: '', recipients: '' });
-				return response.data.entry_id;
-			})
-			.then(async (entry_id) => {
-				if (files.length !== 0) {
-					const file_names = files.map((item) => item.name);
-					try {
-						await getPresignedURL({ attachment_names: file_names }).then(async (res) => {
-							var presigned_urls = { ...res.data.url_dict };
-
-							await awsFileUpload(presigned_urls, files, entry_id).then((response) => {
-								setFiles([]);
-								updatedTicket.thread.entries.at(-1).attachments.push(...response);
-							});
-						});
-					} catch(err) {
-						alert('Attachment will not added to thread. Configure S3 to allow attachments');
-						console.error(err);
-					}
-				}
 			})
 			.then(() => {
 				updateCurrentTicket(updatedTicket);
-			})
-			.catch((err) => {
-				alert('Error creating thread');
-				console.error(err);
 			});
+		}
 	};
 
-	const awsFileUpload = async (presigned_urls, files, entry_id) => {
+	const awsFileUpload = async (presigned_urls, files) => {
 		var attachments = [];
 		await Promise.all(
 			Object.entries(presigned_urls).map(([fileName, url]) => {
 				const file = files.find((f) => f.name === fileName);
-
 				try {
-					return axios
-						.put(url, file, {
-							headers: {
-								'Content-Type': file.type,
-								'Content-Disposition': `inline; filename="${fileName}"`,
-							},
-						})
-						.then(async (res) => {
-							await createAttachment({
-								object_id: entry_id,
-								type: file.type,
-								name: fileName,
-								inline: 1,
-								link: url.split('?')[0],
-								size: file.size,
-							}).then((res) => {
-								attachments.push(res.data);
-							});
-						})
-						.catch((err) => {
-							console.error(`Error creating attachment in db for ${fileName}: `, err);
-						});
+					axios.put(url, file, {
+						headers: {
+							'Content-Type': file.type,
+							'Content-Disposition': `inline; filename="${fileName}"`,
+						},
+					});
+					attachments.push({ name: fileName, link: url.split('?')[0], type: file.type, size: file.size, inline: 1});
 				} catch (error) {
 					console.error(`Error uploading ${fileName}:`, error.message);
 				}
@@ -163,29 +161,25 @@ export const TicketThread = ({ ticket, closeDrawer, updateCurrentTicket, type })
 	const handleFileUpload = (event) => {
 		const length = event.target.files.length;
 		var tempArray = [];
-		var sizeExceed = false
-		for (let i = 0; i < length; i++) {	
+		var sizeExceed = false;
+		for (let i = 0; i < length; i++) {
 			if (event.target.files[i].size > Number(maxUploadSize)) {
-				sizeExceed = true
-				continue
+				sizeExceed = true;
+				continue;
 			}
 			tempArray.push(event.target.files[i]);
 		}
 		setFiles((p) => [...p, ...tempArray]);
 		event.target.value = '';
 
-		if(sizeExceed) {
-			alert(`One or more files exceed the max upload limit of ${humanFileSize(maxUploadSize, true)}!`)
+		if (sizeExceed) {
+			alert(`One or more files exceed the max upload limit of ${humanFileSize(maxUploadSize, true)}!`);
 		}
 	};
 
 	const handleDeleteFile = (idx) => {
 		setFiles((p) => [...p.slice(0, idx), ...p.slice(idx + 1)]);
 	};
-
-	// useEffect(() => {
-	// 	console.log(files);
-	// }, [files]);
 
 	function getEventText(item) {
 		const capitilize = (str) => {
@@ -482,9 +476,7 @@ export const TicketThread = ({ ticket, closeDrawer, updateCurrentTicket, type })
 								<CustomInput type='file' onChange={(event) => handleFileUpload(event)} multiple />
 							</Stack>
 						}
-					>
-
-					</RichTextEditorBox>
+					></RichTextEditorBox>
 
 					{/* <Box mt={2} width={'100%'} textAlign={'center'}>
 						<CircularButton sx={{ py: 2, px: 6, width: 200 }} onClick={handleSubmit} disabled={postDisabled}>
